Fix the automatic mode selection
Problem in commit 16ae930 solved. No need for a maximum number of oscillations,
but the dynamic array storage routines deal with real types, whereas the mode is an
integer. Without proper conversion, it was always set to 0 (cooling mode), hence the
weird behaviour when entering the deadband from heating mode.